+++ modules/loginapi/src/main/java/allcom/service/SmsService.java
@@ -25,33 +25,46 @@ public class SmsService {
 
     @Value("${systemparam.verifycodetimeout}")
     private long verifyCodeTimeout;
+    @Value("${systemparam.verifysmscountlimit}")
+    private int verifySmsCountLimit;
 
     @Autowired
     SmsVerifyCodeRepository smsVerifyCodeRepository;
 
-    public RetMessage sendSms(String phoneNumber,String area){
+    public RetMessage sendSms(String sessionId,int umid,String phoneNumber,String area){
         RetMessage retMessage = new RetMessage();
         //key 为纯数字6位随机串
         String key = GlobalTools.getRandomString(6,true);
         //调用短信发送接口发送短信，传入参数为phoneNumber和key
-        boolean sendresult = true;
+
+        if(umid>0 && (smsVerifyCodeRepository.findSuccessCountByUmid(umid)>verifySmsCountLimit)) {
+            retMessage.setErrorCode("-15");
+            retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area, "-15"));
+            log.info("send sms count exceeds limited number!!! umid is:" + umid);
+        }else if(!sessionId.equals("") && (smsVerifyCodeRepository.findSuccessCountBySessionId(sessionId)>verifySmsCountLimit)){
+            retMessage.setErrorCode("-15");
+            retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area, "-15"));
+            log.info("send sms count exceeds limited number!!! sessionId is:" + sessionId);
+        }else {
+            boolean sendresult = true;
 //        sendresult = ...
 
-        Timestamp sendTime = new Timestamp(System.currentTimeMillis());
-        SmsVerifyCode smsVerifyCode = new SmsVerifyCode(phoneNumber,key,sendTime,"0");
+            Timestamp sendTime = new Timestamp(System.currentTimeMillis());
+            SmsVerifyCode smsVerifyCode = new SmsVerifyCode(sessionId,umid,phoneNumber, key, sendTime, "0");
 
-        if(sendresult){
-            retMessage.setErrorCode("0");
-            retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area,"0"));
-        }else{
-            retMessage.setErrorCode("-6");
-            retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area,"-6"));
-            smsVerifyCode.setSendResult("-6");
+            if (sendresult) {
+                retMessage.setErrorCode("0");
+                retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area, "0"));
+            } else {
+                retMessage.setErrorCode("-6");
+                retMessage.setErrorMessage(GlobalTools.getMessageByLocale(area, "-6"));
+                smsVerifyCode.setSendResult("-6");
+            }
+            smsVerifyCodeRepository.save(smsVerifyCode);
+
+            //删除1天前的短信验证码记录
+            smsVerifyCodeRepository.deleteOldRecord(GlobalTools.getTimeBefore(3600 * 24));
         }
-        smsVerifyCodeRepository.save(smsVerifyCode);
-
-        //删除1天前的短信验证码记录
-        smsVerifyCodeRepository.deleteOldRecord(GlobalTools.getTimeBefore(3600*24));
 
         return retMessage;
     }
